+++ src/rpc/server.ts
@@ -30,12 +30,15 @@ export class RpcServer {
 	}
 
 	/**
-	 * Maps a method handler, allowing the server to receive requests or notifications.
+	 * Adds a method handler, allowing the server to receive requests and notifications.
 	 * @param name The name of the method to be mapped.
 	 * @param handler The handler function to be invoked when the method is called.
 	 * @returns A disposable object that can be used to remove the method handler.
 	 */
-	public add<TParameters extends RpcRequestParameters>(name: string, handler: MethodHandler<TParameters>): IDisposable {
+	public addMethod<TParameters extends RpcRequestParameters>(
+		name: string,
+		handler: MethodHandler<TParameters>,
+	): IDisposable {
 		return this.#methods.disposableOn(name, handler as MethodHandler);
 	}